+++ src-java/flowhs-topology/flowhs-storm-topology/src/main/java/org/openkilda/wfm/topology/flowhs/bolts/HaFlowCreateHubBolt.java
@@ -15,6 +15,7 @@
 
 package org.openkilda.wfm.topology.flowhs.bolts;
 
+import static org.openkilda.wfm.topology.flowhs.FlowHsTopology.Stream.HUB_TO_FLOW_MONITORING_TOPOLOGY_SENDER;
 import static org.openkilda.wfm.topology.flowhs.FlowHsTopology.Stream.HUB_TO_HISTORY_TOPOLOGY_SENDER;
 import static org.openkilda.wfm.topology.flowhs.FlowHsTopology.Stream.HUB_TO_METRICS_BOLT;
 import static org.openkilda.wfm.topology.flowhs.FlowHsTopology.Stream.HUB_TO_NB_RESPONSE_SENDER;
@@ -30,6 +31,7 @@ import org.openkilda.messaging.Message;
 import org.openkilda.messaging.command.CommandData;
 import org.openkilda.messaging.command.haflow.HaFlowRequest;
 import org.openkilda.messaging.info.InfoMessage;
+import org.openkilda.messaging.info.stats.RemoveFlowPathInfo;
 import org.openkilda.messaging.info.stats.UpdateFlowPathInfo;
 import org.openkilda.pce.AvailableNetworkFactory;
 import org.openkilda.pce.PathComputer;
@@ -61,7 +63,7 @@ import org.apache.storm.tuple.Fields;
 import org.apache.storm.tuple.Tuple;
 import org.apache.storm.tuple.Values;
 
-public class HaFlowCreateHubBolt  extends HubBolt implements FlowGenericCarrier {
+public class HaFlowCreateHubBolt extends HubBolt implements FlowGenericCarrier {
     private final HaFlowCreateConfig config;
     private final PathComputerConfig pathComputerConfig;
     private final FlowResourcesConfig flowResourcesConfig;
@@ -154,7 +156,7 @@ public class HaFlowCreateHubBolt  extends HubBolt implements FlowGenericCarrier 
 
     @Override
     public void sendHistoryUpdate(@NonNull FlowHistoryHolder historyHolder) {
-        //TODO check history
+        //TODO check history https://github.com/telstra/open-kilda/issues/5169
         InfoMessage message = new InfoMessage(historyHolder, getCommandContext().getCreateTime(),
                 getCommandContext().getCorrelationId());
         emitWithContext(Stream.HUB_TO_HISTORY_TOPOLOGY_SENDER.name(), getCurrentTuple(),
@@ -163,17 +165,22 @@ public class HaFlowCreateHubBolt  extends HubBolt implements FlowGenericCarrier 
 
     @Override
     public void sendNotifyFlowStats(@NonNull UpdateFlowPathInfo flowPathInfo) {
-        //TODO implement
+        //TODO implement https://github.com/telstra/open-kilda/issues/5182
+    }
+
+    @Override
+    public void sendNotifyFlowStats(RemoveFlowPathInfo flowPathInfo) {
+        //TODO implement https://github.com/telstra/open-kilda/issues/5182
     }
 
     @Override
     public void sendNotifyFlowMonitor(@NonNull CommandData flowCommand) {
-        //TODO implement
+        //TODO implement https://github.com/telstra/open-kilda/issues/5172
     }
 
     @Override
     public void sendPeriodicPingNotification(String haFlowId, boolean enabled) {
-        //TODO implement periodic pings
+        //TODO implement periodic pings https://github.com/telstra/open-kilda/issues/5153
         log.info("Periodic pings are not implemented for ha-flow create operation yet. Skipping for the ha-flow {}",
                 haFlowId);
     }
@@ -196,6 +203,7 @@ public class HaFlowCreateHubBolt  extends HubBolt implements FlowGenericCarrier 
         declarer.declareStream(HUB_TO_NB_RESPONSE_SENDER.name(), MessageKafkaTranslator.STREAM_FIELDS);
         declarer.declareStream(HUB_TO_HISTORY_TOPOLOGY_SENDER.name(), MessageKafkaTranslator.STREAM_FIELDS);
         declarer.declareStream(HUB_TO_PING_SENDER.name(), MessageKafkaTranslator.STREAM_FIELDS);
+        declarer.declareStream(HUB_TO_FLOW_MONITORING_TOPOLOGY_SENDER.name(), MessageKafkaTranslator.STREAM_FIELDS);
         declarer.declareStream(ZkStreams.ZK.toString(),
                 new Fields(ZooKeeperBolt.FIELD_ID_STATE, ZooKeeperBolt.FIELD_ID_CONTEXT));
     }