+++ src-java/testing/test-library/src/main/java/org/openkilda/testing/service/northbound/NorthboundServiceV2.java
@@ -32,6 +32,7 @@ import org.openkilda.northbound.dto.v2.flows.PathValidateResponse;
 import org.openkilda.northbound.dto.v2.haflows.HaFlow;
 import org.openkilda.northbound.dto.v2.haflows.HaFlowCreatePayload;
 import org.openkilda.northbound.dto.v2.haflows.HaFlowPatchPayload;
+import org.openkilda.northbound.dto.v2.haflows.HaFlowPaths;
 import org.openkilda.northbound.dto.v2.haflows.HaFlowUpdatePayload;
 import org.openkilda.northbound.dto.v2.links.BfdProperties;
 import org.openkilda.northbound.dto.v2.links.BfdPropertiesPayload;
@@ -196,4 +197,6 @@ public interface NorthboundServiceV2 {
     HaFlow partialUpdateHaFlow(String haFlowId, HaFlowPatchPayload request);
 
     HaFlow deleteHaFlow(String haFlowId);
+
+    HaFlowPaths getHaFlowPaths(String haFlowId);
 }